Enhance .gitignore for better IDE support
This commit modifies the .gitignore file, it now additionally covers `.idea/` directory to increase support for IntelliJ IDEA which is widely used in the developer community besides other IDEs. By ignoring `.idea/`, we ensure that user-specific settings do not get pushed to the repository, thereby maintaining a clean and consistent codebase for all contributors.