Refactor append operation and add ToJSON method
Refactored the "printEach" method in "methods.go" to reduce code repetition, by extracting the append operation to a new method named "appendEach".
Also, added the "ToJSON" method to "entities.go" to drive conversion of each model entity to a JSON representation. This changeset allows for cleaner code and an easier way to convert model entities to JSON.