Add JSON serialization and save to file functionality
Added functionality to serialize models into JSON format and save it to a file. The methods.go file was modified to include functions to convert models to JSON and write to a file. Also, the .gitignore file was changed to ignore generated JSON files, to prevent them from being committed.